// app/src/interfaces/exampleInterfaces.tsx
export interface UseExample {
    id_example: number;
    correo_example: string;
    nota_example: string;
    ciudad_example: string;
    proveedor_example: string;
    image1_example: string;
    image2_example: string;
    image3_example: string;
    prioridad_example: number;
    updaate: string;
}

export type RequestLogin = UseExample | false;